+++ src/datasources/data-frame/DataFrameDataSourceBase.ts
@@ -16,7 +16,8 @@ import {
     DataFrameDataQuery,
     DataFrameVariableQuery,
     Option,
-    ColumnFilter
+    ColumnFilter,
+    CombinedFilters
 } from './types';
 import { BackendSrv, TemplateSrv } from '@grafana/runtime';
 import { extractErrorInfo } from 'core/errors';
@@ -63,7 +64,7 @@ export abstract class DataFrameDataSourceBase<
     ): Promise<TableDataRows>;
 
     public abstract queryTables$(
-        query: string,
+        filters: CombinedFilters,
         take?: number,
         projection?: DataTableProjections[]
     ): Observable<TableProperties[]>;